+++ src/mantle/testing/class-test-case.php
@@ -92,6 +92,13 @@ abstract class Test_Case extends BaseTestCase {
 	 * Runs the routine before setting up all tests.
 	 */
 	public static function setUpBeforeClass(): void {
+		parent::setUpBeforeClass();
+
+		// Disable the Spatie once cache for tests.
+		if ( class_exists( \Spatie\Once\Once::class ) ) {
+			\Spatie\Once\Cache::getInstance()->disable();
+		}
+
 		static::register_traits();
 
 		if ( ! empty( static::$test_uses ) ) {