+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	the home URL.
 - Added `with_option()`/`with_home_url()`/`with_site_url()` methods to the installation manager.
 
+### Changed
+
+- Disable `spatie/once`'s cache if found during unit testing.
+
 ## v1.3.3 - 2025-01-10
 
 ### Added